First step towards auto-extracting client
client.py will pull from github, as requested by the server
client.py will extract and put all files locally ( except replacing itself ? )

Moved cutechess directly into the Client folder to it gets pulled.
May need to chmod +x the cutechess-ob for Linux on demand.

In the future, the cleanup script will delete everythinge except:
client.py, and any credentials.* files.
Such that one can run an openbench client in any directory, from a single file.
